# open-gestionale/api/openapi.yaml
openapi: 3.0.3
info:
  title: Open Gestionale API
  version: "1.0"
  description: >
    Contratto REST tra un server Open Roleplay e il suo gestionale web.
    Read-mostly: il bridge legge i dati di gioco e scrive solo su tabelle
    proprie. Ogni risposta è incapsulata in `{ data, meta }`; gli errori in
    `{ error: { code, message } }`. Se un modulo non è attivo, l'endpoint del
    widget risponde 404 con codice `WIDGET_NOT_AVAILABLE`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.gestionale.tuoserver.it
    description: Bridge sul VPS del server
  - url: http://localhost:3000
    description: Sviluppo locale
...
paths:
...
  /widget/{widgetId}:
    get:
      tags: [widget]
      summary: Dati di un widget per il giocatore corrente
      description: >
        widgetId è l'id di registro, es. `economy.balance`, `economy.transactions`,
        `company.detail`, `justice.dossiers`, `politics.laws`. Alcuni widget
        accettano query: `page`, `limit` (collezioni), `id` (`company.detail`),
        `status` (`politics.laws`), `unread` (`notifications.feed`).
      parameters:
        - name: widgetId
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200 }
        - name: id
          in: query
          schema: { type: string }
      responses:
        "200": { description: Payload del widget }
        "403":
          description: Capability insufficiente (WIDGET_FORBIDDEN)
        "404":
          description: Widget sconosciuto o modulo non attivo (WIDGET_NOT_AVAILABLE)
  /widget/notifications.feed/{id}/read:
    post:
      tags: [widget]
      summary: Segna una notifica come letta
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/Error" }
  /widget/economy.transfers:
    post:
      tags: [widget]
      summary: Bonifico (unica scrittura, abilitata solo in demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, amount]
              properties:
                to: { type: string }
                amount: { type: number }
                note: { type: string }
      responses:
        "200": { description: Bonifico simulato }
        "403": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Errore
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: WIDGET_NOT_AVAILABLE }
            message: { type: string }
security:
  - bearerAuth: []